fix: remove `element.ref` warning
