+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@sanity/ui",
-  "version": "2.11.6",
+  "version": "2.11.7-canary.0",
   "keywords": [
     "sanity",
     "ui",
